New class Normalizer. moved related methods from SearchEngine to Normalizer
